requirements.txt
streamlit==1.32.0
snowflake-snowpark-python==1.7.0
pandas==2.2.1
plotly==5.19.0
python-dotenv==1.0.1 